Create server-transactions types
