+++ src/components/TransactionsInterpreter/helpers/tests/base-transaction-mock.ts
@@ -1,4 +1,4 @@
-import { TxActionsEnum } from '../types';
+import { TxActionsEnum } from 'types/server-transations';
 
 export const baseTransactionMock = {
   blockHash: '',
