+++ src/components/TransactionsInterpreter/helpers/tests/getTransactionMethod.test.ts
@@ -1,5 +1,9 @@
+import {
+  TransactionType,
+  TxActionCategoryEnum,
+  TxActionsEnum
+} from 'types/server-transations';
 import { getTransactionMethod } from '../getTransactionMethod';
-import { TransactionType, TxActionCategoryEnum, TxActionsEnum } from '../types';
 import { baseTransactionMock } from './base-transaction-mock';
 
 describe('getTransactionMethod', () => {
