+++ src/components/TransactionsInterpreter/helpers/tests/getTransactionReceiver.test.ts
@@ -1,5 +1,9 @@
+import {
+  TransactionType,
+  TxActionCategoryEnum,
+  TxActionsEnum
+} from 'types/server-transations';
 import { getTransactionReceiver } from '../getTransactionReceiver';
-import { TransactionType, TxActionCategoryEnum, TxActionsEnum } from '../types';
 import { baseTransactionMock } from './base-transaction-mock';
 
 describe('getTransactionReceiver', () => {
