+++ src/types/index.ts
@@ -16,6 +16,7 @@ export interface TokenLoginType {
 }
 
 export * from './transactions';
+export * from './server-transations';
 export * from './enums';
 export * from './account';
 export * from './network';
